Downgrade Logstash for Monasca
Logstash was added to Kolla as part of the effort to deploy Monasca.

Monasca uses Kafka messaging version 0.9 which is not supported by
Logstash 5.x with the default Kafka input/output plugins. An attempt
to downgrade the plugins to the 4.x series which should be compatible
with messaging version 0.9 and Logstash 5.x failed. Therefore, until
Monasca is upgraded (work is ongoing) this patch downgrades
Logstash to 2.4.1.

Although it would have been simpler to perform the downgrade by
blacklisting the Logstash 5.x package which is provided as part of the
official ELK repo and installing the official repo for Logstash 2.4.1,
Ubuntu 18.04 no longer accepts SHA1 hashes (for good reason) and will
not install Logstash 2.4.1 from this repo (fails with GPG error).
This approach does however work for Centos, but for consistency it
was not used.

This change therefore modifies the Logstash dockerfile, and directly
downloads and installs the Logstash 2.4.1 package. It should be
reverted as soon as Monasca is upgraded.

Partially-Implements: blueprint monasca-containers
Change-Id: I58fba9ceded4b3a198136df084c7d37365ab47a8